requirements.txt
# L21 - Iris Classification using Naive Bayes
# Python 3.10+ required

# Core scientific computing
numpy>=1.24.0

# Data manipulation
pandas>=2.0.0

# Machine learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration management
pyyaml>=6.0
